+++ src/client/components/Dashboard/my-tasks/constants.js
@@ -1,12 +1,3 @@
-export const companyOptions = (companies) => {
-  const companiesList = companies.map((company) => ({
-    label: company.name,
-    value: company.id,
-  }))
-
-  return [SHOW_ALL_OPTION, ...companiesList]
-}
-
 const ME_OTHERS_OPTIONS = [
   {
     label: 'Me',
