+++ src/client/components/Dashboard/my-tasks/transformers.js
@@ -1,0 +1,10 @@
+import { SHOW_ALL_OPTION } from './constants'
+
+export const companyOptions = (companies) => {
+  const companiesList = companies.map((company) => ({
+    label: company.name,
+    value: company.id,
+  }))
+
+  return [SHOW_ALL_OPTION, ...companiesList]
+}
